+++ src/ll/orasql/scripts/oradrop.py
@@ -9,7 +9,7 @@
 ## See orasql/__init__.py for the license
 
 
-import sys, os, optparse
+import sys, os, argparse
 
 from ll import astyle, orasql
 
@@ -21,25 +21,21 @@ s4object = astyle.Style.fromenv("LL_ORASQL_REPRANSI_OBJECT", "green:black")
 
 
 def main(args=None):
-	colors = ("yes", "no", "auto")
-	fks = ("keep", "disable", "drop")
-	p = optparse.OptionParser(usage="usage: %prog [options] connectstring >output.sql")
-	p.add_option("-v", "--verbose", dest="verbose", help="Give a progress report?", default=False, action="store_true")
-	p.add_option("-c", "--color", dest="color", help="Color output ({0})".format(", ".join(colors)), default="auto", choices=colors)
-	p.add_option("-f", "--fks", dest="fks", help="How should foreign keys from other schemas be treated ({0})?".format(", ".join(fks)), default="disable", choices=fks)
-	p.add_option("-x", "--execute", dest="execute", action="store_true", help="immediately execute the commands instead of printing them?")
-	p.add_option("-k", "--keepjunk", dest="keepjunk", help="Output objects with '$' in their name?", default=False, action="store_true")
-	p.add_option("-i", "--ignore", dest="ignore", help="Ignore errors?", default=False, action="store_true")
-	p.add_option("-e", "--encoding", dest="encoding", help="Encoding for output", default="utf-8")
+	p = argparse.ArgumentParser(description="Print (or execute) drop statements for all objects in an Oracle database schema")
+	p.add_argument("connectstring", help="Oracle connect string")
+	p.add_argument("-v", "--verbose", dest="verbose", help="Give a progress report?", default=False, action="store_true")
+	p.add_argument("-c", "--color", dest="color", help="Color output", default="auto", choices=("yes", "no", "auto"))
+	p.add_argument("-f", "--fks", dest="fks", help="How should foreign keys from other schemas be treated?", default="disable", choices=("keep", "disable", "drop"))
+	p.add_argument("-x", "--execute", dest="execute", action="store_true", help="immediately execute the commands instead of printing them?")
+	p.add_argument("-k", "--keepjunk", dest="keepjunk", help="Output objects with '$' in their name?", default=False, action="store_true")
+	p.add_argument("-i", "--ignore", dest="ignore", help="Ignore errors?", default=False, action="store_true")
+	p.add_argument("-e", "--encoding", dest="encoding", help="Encoding for output", default="utf-8")
 
-	(options, args) = p.parse_args(args)
-	if len(args) != 1:
-		p.error("incorrect number of arguments")
-		return 1
+	args = p.parse_args(args)
 
-	if options.color == "yes":
+	if args.color == "yes":
 		color = True
-	elif options.color == "no":
+	elif args.color == "no":
 		color = False
 	else:
 		color = None
@@ -47,16 +43,16 @@ def main(args=None):
 	stdout = astyle.Stream(sys.stdout, color)
 	stderr = astyle.Stream(sys.stderr, color)
 
-	connection = orasql.connect(args[0])
+	connection = orasql.connect(args.connectstring)
 
-	term = not options.execute
+	term = not args.execute
 
 	cs = s4connectstring(connection.connectstring())
 
 	def keep(obj):
 		if obj.owner is not None and not isinstance(obj, orasql.ForeignKey):
 			return False
-		if options.keepjunk:
+		if args.keepjunk:
 			return True
 		if "$" in obj.name:
 			return False
@@ -70,10 +66,10 @@ def main(args=None):
 		action = "skipped"
 		if obj.owner is not None:
 			if isinstance(obj, orasql.ForeignKey):
-				if options.fks == "disable":
+				if args.fks == "disable":
 					ddl = obj.disableddl(cursor, term)
 					action = "disabled"
-				elif options.fks == "drop":
+				elif args.fks == "drop":
 					ddl = obj.dropddl(cursor, term)
 					action = None
 		elif keepdef:
@@ -81,7 +77,7 @@ def main(args=None):
 			action = None
 
 		# Progress report
-		if options.verbose:
+		if args.verbose:
 			msg = astyle.style_default("oradrop.py: ", cs, ": fetching #{0} ".format(i+1), s4object(str(obj)))
 			if action is not None:
 				msg = astyle.style_default(msg, " ", s4warning("({0})".format(action)))
@@ -89,21 +85,21 @@ def main(args=None):
 
 		if ddl:
 			# Print or execute DDL
-			if options.execute:
+			if args.execute:
 				ddls.append((obj, ddl))
 			else:
-				stdout.write(ddl.encode(options.encoding))
+				stdout.write(ddl.encode(args.encoding))
 
 	# Execute DDL
-	if options.execute:
+	if args.execute:
 		cursor = connection.cursor()
 		for (i, (obj, ddl)) in enumerate(ddls):
-			if options.verbose:
+			if args.verbose:
 				stderr.writeln("oradrop.py: ", cs, ": dropping #{0}/{1} ".format(i+1, len(ddls)), s4object(str(obj)))
 			try:
 				cursor.execute(ddl)
 			except orasql.DatabaseError, exc:
-				if not options.ignore or "ORA-01013" in str(exc):
+				if not args.ignore or "ORA-01013" in str(exc):
 					raise
 				stderr.writeln("oradrop.py: ", s4error("{0}: {1}".format(exc.__class__, str(exc).strip())))