+++ src/ll/scripts/ucat.py
@@ -10,7 +10,7 @@
 ## See ll/__init__.py for the license
 
 
-import sys, optparse, contextlib, errno
+import sys, argparse, contextlib, errno
 
 from ll import url
 
@@ -28,7 +28,7 @@ except ImportError:
 def main(args=None):
 	def catone(urlread):
 		if urlread.isdir():
-			if options.recursive:
+			if args.recursive:
 				for u in urlread.listdir():
 					catone(urlread/u)
 			else:
@@ -44,23 +44,19 @@ def main(args=None):
 						else:
 							break
 			except Exception:
-				if not options.ignoreerrors:
+				if not args.ignoreerrors:
 					raise
 
-	colors = ("yes", "no", "auto")
-	p = optparse.OptionParser(usage="usage: %prog [options] source-file-url target-file-url\n   or: %prog [options] source-file-url(s) target-dir-url")
-	p.add_option("-v", "--verbose", dest="verbose", help="Be verbose?", action="store_true", default=False)
-	p.add_option("-r", "--recursive", dest="recursive", help="Copy stuff recursively?", action="store_true", default=False)
-	p.add_option("-x", "--ignoreerrors", dest="ignoreerrors", help="Ignore errors?", action="store_true", default=False)
+	p = argparse.ArgumentParser(description="print URL content on the screen")
+	p.add_argument("urls", metavar="url", help="URLs to be printed", nargs="+", type=url.URL)
+	p.add_argument("-v", "--verbose", dest="verbose", help="Be verbose?", action="store_true", default=False)
+	p.add_argument("-r", "--recursive", dest="recursive", help="Copy stuff recursively?", action="store_true", default=False)
+	p.add_argument("-x", "--ignoreerrors", dest="ignoreerrors", help="Ignore errors?", action="store_true", default=False)
 
-	(options, args) = p.parse_args(args)
-	if len(args) < 1:
-		p.error("need at least one url")
-		return 1
-
+	args = p.parse_args(args)
 	with url.Context():
-		for arg in args:
-			catone(url.URL(arg))
+		for u in args.urls:
+			catone(u)
 
 
 if __name__ == "__main__":