+++ src/ll/scripts/ucp.py
@@ -10,7 +10,7 @@
 ## See ll/__init__.py for the license
 
 
-import sys, optparse, contextlib
+import sys, argparse, contextlib
 
 from ll import url
 
@@ -28,15 +28,15 @@ except ImportError:
 def main(args=None):
 	def copyone(urlread, urlwrite):
 		if urlread.isdir():
-			if options.recursive:
+			if args.recursive:
 				for u in urlread.listdir():
 					copyone(urlread/u, urlwrite/u)
 			else:
-				if options.verbose:
+				if args.verbose:
 					msg = astyle.style_default("ucp: ", astyle.style_url(str(urlread)), " (directory skipped)")
 					stderr.writeln(msg)
 		else:
-			if options.verbose:
+			if args.verbose:
 				msg = astyle.style_default("ucp: ", astyle.style_url(str(urlread)), " -> ")
 				stderr.write(msg)
 			try:
@@ -53,61 +53,61 @@ def main(args=None):
 				if user or group:
 					urlwrite.chown(user, group)
 			except Exception:
-				if options.ignoreerrors:
-					if options.verbose:
+				if args.ignoreerrors:
+					if args.verbose:
 						msg = astyle.style_error(" (failed)")
 						stderr.writeln(msg)
 				else:
 					raise
 			else:
-				if options.verbose:
+				if args.verbose:
 					msg = astyle.style_default(astyle.style_url(str(urlwrite)), " (", str(size), " bytes)")
 					stderr.writeln(msg)
 
 
 
-	colors = ("yes", "no", "auto")
-	p = optparse.OptionParser(usage="usage: %prog [options] source-file-url target-file-url\n   or: %prog [options] source-file-url(s) target-dir-url")
-	p.add_option("-v", "--verbose", dest="verbose", help="Be verbose?", action="store_true", default=False)
-	p.add_option("-c", "--color", dest="color", help="Color output ({0})".format(", ".join(colors)), default="auto", choices=colors)
-	p.add_option("-u", "--user", dest="user", help="user id or name for target files")
-	p.add_option("-g", "--group", dest="group", help="group id or name for target files")
-	p.add_option("-r", "--recursive", dest="recursive", help="Copy stuff recursively?", action="store_true", default=False)
-	p.add_option("-x", "--ignoreerrors", dest="ignoreerrors", help="Ignore errors?", action="store_true", default=False)
+	p = argparse.ArgumentParser(description="Copies URLs")
+	p.add_argument("urls", metavar="url", help="either one source and one target file, or multiple source files and one target dir", nargs="*", type=url.URL)
+	p.add_argument("-v", "--verbose", dest="verbose", help="Be verbose?", action="store_true", default=False)
+	p.add_argument("-c", "--color", dest="color", help="Color output", default="auto", choices=("yes", "no", "auto"))
+	p.add_argument("-u", "--user", dest="user", help="user id or name for target files")
+	p.add_argument("-g", "--group", dest="group", help="group id or name for target files")
+	p.add_argument("-r", "--recursive", dest="recursive", help="Copy stuff recursively?", action="store_true", default=False)
+	p.add_argument("-x", "--ignoreerrors", dest="ignoreerrors", help="Ignore errors?", action="store_true", default=False)
 
-	(options, args) = p.parse_args(args)
-	if len(args) < 2:
+	args = p.parse_args(args)
+	if len(args.urls) < 2:
 		p.error("need at least one source url and one target url")
 		return 1
 
-	if options.color == "yes":
+	if args.color == "yes":
 		color = True
-	elif options.color == "no":
+	elif args.color == "no":
 		color = False
 	else:
 		color = None
 	stdout = astyle.Stream(sys.stdout, color)
 	stderr = astyle.Stream(sys.stderr, color)
 
-	user = options.user
+	user = args.user
 	try:
 		user = int(user)
 	except (TypeError, ValueError):
 		pass
 
-	group = options.group
+	group = args.group
 	try:
 		group = int(group)
 	except (TypeError, ValueError):
 		pass
 
 	with url.Context():
-		args = [url.URL(arg) for arg in args]
-		if len(args) > 2 or args[-1].isdir(): # treat target as directory
-			for arg in args[:-1]:
-				copyone(arg, args[-1]/arg.file)
+		urls = args.urls
+		if len(urls) > 2 or urls[-1].isdir(): # treat target as directory
+			for u in urls[:-1]:
+				copyone(u, urls[-1]/u.file)
 		else:
-			copyone(args[0], args[-1])
+			copyone(urls[0], urls[-1])
 
 
 if __name__ == "__main__":